# openapi.yml
openapi: 3.0.3
info:
  title: Office Commute API
  description: |
    사내 출퇴근/근태/연차/조직 관리 시스템 API.
    모든 응답 오류는 `ErrorResult { code, message }` 또는
    `ValidationErrorResult { code, message, fieldErrorResults }` 형식을 따른다.
  version: 0.0.1
servers:
  - url: http://localhost:8080
    description: 로컬 개발
...
security:
  - SessionAuth: []
paths:
...
  /team:
    post:
      tags: [Team]
      summary: 팀 등록 (Manager Only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TeamRegisterRequest' }
      responses:
        '201':
          description: 등록 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TeamRegisterResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: 동일 이름의 팀이 이미 존재 (TEAM_ALREADY_EXISTS)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResult' }
    get:
      tags: [Team]
      summary: 팀 전체 조회
      description: |
        `memberCount`는 저장된 카운터가 아니라 `employee.team_id`의 COUNT 결과를 응답 시점에 합성한 값이다.
      responses:
        '200':
          description: 팀 목록
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TeamFindResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
...
  responses:
...
    ValidationError:
      description: 입력값 검증 실패 (VALIDATION_ERROR)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationErrorResult' }
    Unauthorized:
      description: 비로그인 (UNAUTHORIZED)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResult' }
    Forbidden:
      description: 권한 부족 (FORBIDDEN)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResult' }
  schemas:
...
    ErrorResult:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            에러 코드. 예시:
            - BAD_REQUEST, INVALID_JSON, INVALID_PARAMETER, MISSING_PARAMETER
            - UNAUTHORIZED, FORBIDDEN
            - TEAM_ALREADY_EXISTS, EMPLOYEE_ALREADY_EXISTS
            - TEAM_NOT_FOUND, EMPLOYEE_NOT_FOUND
            - DUPLICATE_WORK, DATA_INTEGRITY_ERROR
            - HOLIDAY_DATA_UNAVAILABLE
            - INTERNAL_SERVER_ERROR
        message: { type: string }
    FieldErrorResult:
      type: object
      required: [field, message]
      properties:
        field: { type: string }
        message: { type: string }
    ValidationErrorResult:
      type: object
      required: [code, message, fieldErrorResults]
      properties:
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string, example: 입력값이 올바르지 않습니다 }
        fieldErrorResults:
          type: array
          items: { $ref: '#/components/schemas/FieldErrorResult' }
...
    TeamRegisterRequest:
      type: object
      required: [teamName]
      properties:
        teamName: { type: string }
        managerName:
          type: string
          nullable: true
          description: 옵셔널. 빈 문자열은 null로 정규화됨.
        annualLeaveCriteria:
          type: integer
          format: int32
          minimum: 0
          nullable: true
          default: 0
          description: 팀 연차 등록 기준. null 또는 생략 시 0으로 처리됨.
    TeamRegisterResponse:
      type: object
      required: [teamId]
      properties:
        teamId: { type: integer, format: int64 }
    TeamFindResponse:
      type: object
      required: [teamId, name, annualLeaveCriteria, memberCount]
      properties:
        teamId: { type: integer, format: int64 }
        name: { type: string }
        managerName: { type: string, nullable: true }
        annualLeaveCriteria:
          type: integer
          format: int32
          minimum: 0
        memberCount:
          type: integer
          format: int64
          description: employee.team_id COUNT 파생값.

    # ---- Employee ----
...